requirements.txt
beautifulsoup4
black
docling
litellm
neo4j
pydantic
python-dotenv
tqdm